Refactor ToolContentModelHooks => ToolModelHooks
