+++ src/models/tools/image/image-registration.ts
@@ -1,5 +1,5 @@
 import { registerToolContentInfo } from "../tool-content-info";
-import { ToolMetadataModel } from "../tool-types";
+import { ToolMetadataModel } from "../tool-metadata";
 import { kImageToolID, ImageContentModel, defaultImageContent } from "./image-content";
 import ImageToolComponent from "../../../components/tools/image-tool";
 import ImageToolIcon from "../../../clue/assets/icons/image-tool.svg";
