+++ src/plugins/data-card-tool/data-card-registration.ts
@@ -1,9 +1,9 @@
 import { registerToolContentInfo } from "../../models/tools/tool-content-info";
+import { ToolMetadataModel } from "../../models/tools/tool-metadata";
 import { kDataCardDefaultHeight, kDataCardToolID } from "./data-card-types";
 import DataCardToolIcon from "./assets/data-card-tool.svg";
 import { DataCardToolComponent } from "./data-card-tool";
 import { defaultDataCardContent, DataCardContentModel } from "./data-card-content";
-import { ToolMetadataModel } from "../../models/tools/tool-types";
 
 
 registerToolContentInfo({
