+++ src/plugins/dataflow-tool/dataflow-registration.ts
@@ -2,9 +2,9 @@ import {
   DataflowContentModel, defaultDataflowContent, kDataflowDefaultHeight, kDataflowToolID
 } from "./model/dataflow-content";
 import { registerToolContentInfo } from "../../models/tools/tool-content-info";
+import { ToolMetadataModel } from "../../models/tools/tool-metadata";
 import DataflowToolComponent from "./components/dataflow-tool";
 import DataflowToolIcon from "./assets/program.svg";
-import { ToolMetadataModel } from "../../../src/models/tools/tool-types";
 
 registerToolContentInfo({
   id: kDataflowToolID,
